Store code refactor
