+++ optaweb-tsp-ui/src/index.old.js
@@ -19,7 +19,7 @@ import * as ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './App';
 import './index.css';
-import configureStore from './store';
+import configureStore from './store/configStore';
 
 import registerServiceWorker from './registerServiceWorker';
